move tests around since tests no longer internal
